# storegateway: remove chunk ref in `AggrChunk` type