… pkg/storegateway/storepb/types.pb.go