<!--  Thanks for sending a pull request!  Before submitting:

1. Read our CONTRIBUTING.md guide
2. Rebase your PR if it gets out of sync with main
-->

#### What this PR does

follow-up PR for https://github.com/grafana/mimir/pull/6190

Another `~10%` of total allocations in the `storegateway` component it is due to the initialization of the raw chunk field inside the `AggrChunk` type. 

![Screenshot 2023-10-06 at 09 52 05](https://github.com/grafana/mimir/assets/888899/4ee23937-03a8-44a1-a5a7-857dc9851609)

![Screenshot 2023-10-06 at 09 52 20](https://github.com/grafana/mimir/assets/888899/24407db4-f74d-4e87-a832-1e41e716b2e0)

However, I haven't found any place in the code where it is checked whether this field is null, except during initialization.

This PR marks the field as non-nullable in order to avoid unnecessary heap allocation every time a new chunk is instantiated.

Below I include the benchmark comparation table for `BucketChunkReader_loadChunks` benchmark. 

```
name                                           old time/op    new time/op    delta
BucketChunkReader_loadChunks/blocks-1000-8        868µs ± 6%     879µs ± 2%    ~     (p=0.222 n=9+9)
BucketChunkReader_loadChunks/blocks-10000-8      10.1ms ±18%     9.0ms ± 1%    ~     (p=0.122 n=10+8)
BucketChunkReader_loadChunks/blocks-100000-8      104ms ±12%     100ms ± 4%    ~     (p=0.247 n=10+10)
BucketChunkReader_loadChunks/blocks-1000000-8     1.02s ± 8%     1.00s ± 3%    ~     (p=0.156 n=10+9)

name                                           old alloc/op   new alloc/op   delta
BucketChunkReader_loadChunks/blocks-1000-8       1.28MB ± 1%    1.28MB ± 1%  -0.36%  (p=0.029 n=10+10)
BucketChunkReader_loadChunks/blocks-10000-8      14.0MB ± 6%    13.6MB ± 3%  -3.00%  (p=0.023 n=10+10)
BucketChunkReader_loadChunks/blocks-100000-8      152MB ± 7%     150MB ± 5%    ~     (p=0.730 n=9+9)
BucketChunkReader_loadChunks/blocks-1000000-8    1.57GB ± 6%    1.58GB ±16%    ~     (p=0.780 n=9+10)

name                                           old allocs/op  new allocs/op  delta
BucketChunkReader_loadChunks/blocks-1000-8        10.7k ± 0%      9.7k ± 0%  -9.35%  (p=0.000 n=10+10)
BucketChunkReader_loadChunks/blocks-10000-8        107k ± 0%       97k ± 0%  -9.41%  (p=0.000 n=10+9)
BucketChunkReader_loadChunks/blocks-100000-8      1.07M ± 0%     0.97M ± 0%  -9.33%  (p=0.000 n=9+9)
BucketChunkReader_loadChunks/blocks-1000000-8     10.7M ± 0%      9.7M ± 0%  -9.40%  (p=0.000 n=9+9)
```

#### Which issue(s) this PR fixes or relates to

Fixes N/A