@dimitarvdimitrov 2023-10-06
I'm curious how an old querier might work with a new store-gateway and vice-versa. Maybe it's enough to check the docs on that proto annotation 🤷