@ortuman 2023-10-19
FTR, i've been running a custom image based on this branch for a whole day in one of our dev environments and haven't noticed anything unusual. 

OTOH, taking a look at the diff of the generated proto file, the major change seems to be related to the new field marshalling, and the only noticeable difference is the null check thas has been removed. 

https://github.com/grafana/mimir/pull/6247/files#diff-334406690f1e2d97deb9bf1ff47a2a530836ab955d4dfab9fddbad2b8bb0cecfL1188-L1198

So I understand that this shouldn't make any difference for the querier.